Fixes https://java.net/jira/browse/JAVASERVERFACES-3666, Migrate test/agnostic/vdl/facelets/contracts/viewroot to test/servlet30/contractUsingViewRoot
git-svn-id: https://svn.java.net/svn/mojarra~svn/trunk@14164 761efcf2-d34d-6b61-9145-99dcacc3edf1
