+++ test/servlet30/contractUsingViewRoot/src/main/java/com/sun/faces/test/servlet30/contractusingviewroot/ViewRootContractsELBean.java
@@ -37,34 +37,22 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-package com.sun.faces.test.agnostic.vdl.facelets.contracts.viewroot;
+package com.sun.faces.test.servlet30.contractusingviewroot;
 
-import com.gargoylesoftware.htmlunit.WebClient;
-import com.gargoylesoftware.htmlunit.html.HtmlPage;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import static org.junit.Assert.*;
+import javax.faces.bean.ManagedBean;
+import javax.faces.bean.RequestScoped;
 
-public class Issue2736IT {
+@RequestScoped
+@ManagedBean(name = "viewRootConstractsELBean")
+public class ViewRootContractsELBean {
 
-    private String webUrl;
-    private WebClient webClient;
+    private String contracts = null;
 
-    @Before
-    public void setUp() {
-        webUrl = System.getProperty("integration.url");
-        webClient = new WebClient();
+    public String getContracts() {
+        return contracts;
     }
 
-    @After
-    public void tearDown() {
-        webClient.closeAllWindows();
-    }
-
-    @Test
-    public void testViewRootConstractsEL() throws Exception {
-        HtmlPage page = webClient.getPage(webUrl + "faces/viewRootContractsEL.xhtml");
-        assertEquals(200, page.getWebResponse().getStatusCode());
+    public void setContracts(String contracts) {
+        this.contracts = contracts;
     }
 }
