+++ test/servlet30/contractUsingViewRoot/src/main/webapp/WEB-INF/faces-config.xml
