+++ test/servlet30/contractUsingViewRoot/src/main/webapp/WEB-INF/glassfish-web.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE glassfish-web-app PUBLIC "-//GlassFish.org//DTD GlassFish Application Server 3.1 Servlet 3.0//EN" "http://glassfish.org/dtds/glassfish-web-app_3_0-1.dtd">
 <glassfish-web-app error-url="">
-  <context-root>/test-agnostic-vdl-facelets-contracts-viewroot</context-root>
+  <context-root>/test-servlet30-contractUsingViewRoot</context-root>
   <class-loader delegate="true"/>
   <jsp-config>
     <property name="keepgenerated" value="true">
