+++ test/servlet30/contractUsingViewRoot/src/main/webapp/WEB-INF/web.xml
