+++ test/servlet30/contractUsingViewRoot/src/main/webapp/contracts/layout2/template.xhtml
