+++ test/servlet30/contractUsingViewRoot/src/main/webapp/index.xhtml
