+++ test/servlet30/contractUsingViewRoot/src/main/webapp/template.xhtml
