+++ test/servlet30/pom.xml
@@ -63,6 +63,7 @@
     </dependencies>
     <modules>
         <module>ajax</module>
+        <module>contractUsingViewRoot</module>
         <module>component</module>
         <module>composite</module>
         <module>custom-viewroot-viewscope</module>
